requirements.txt
fastapi
uvicorn
pydantic
pypdf
huggingface_hub
python-multipart
python-dotenv
sqlalchemy
passlib[argon2]
python-jose
email-validator
